:truck: RECORD 그룹에 해당하는 모든 뱃지를 조회하는 테스트를 BadgeRepositoryTest로 이동
